Move InfoField, fix prop error
